+++ kwarray/distributions.py
@@ -28,7 +28,6 @@ import fractions  # NOQA
 from kwarray.util_random import ensure_rng
 
 inf = float('inf')
-# __all__ =  [
 
 
 class Value(ub.NiceRepr):
@@ -679,10 +678,12 @@ class Distribution(Parameterized, _RBinOpMixin):
         Example:
             >>> from kwarray.distributions import Normal  # NOQA
             >>> self = Normal()
+            >>> # xdoctest: +REQUIRES(module:kwplot)
             >>> # xdoctest: +REQUIRES(--show)
             >>> import kwplot
             >>> kwplot.autompl()
             >>> self.plot(n=1000)
+            >>> kwplot.show_if_requested()
         """
         import seaborn as sns
         if isinstance(n, str):
@@ -802,19 +803,24 @@ class Mixture(MixedDistribution):
         >>> # Given two normal distributions,
         >>> from kwarray.distributions import Normal  # NOQA
         >>> from kwarray.distributions import *  # NOQA
-        >>> n1 = Normal(mean=11, std=3)
-        >>> n2 = Normal(mean=53, std=5)
+        >>> n1 = Normal(mean=11, std=3, rng=0)
+        >>> n2 = Normal(mean=53, std=5, rng=0)
         >>> composed = (n1 * 0.3) + (n2 * 0.7)
-        >>> mixture = Mixture([n1, n2], [0.3, 0.7])
+        >>> mixture = Mixture([n1, n2], [0.3, 0.7], rng=0)
+        >>> sample_scalar = mixture.sample()
+        >>> sample_vector = mixture.sample(2)
+        >>> sample_matrix = mixture.sample(2, 3)
+        >>> assert sample_scalar.shape == tuple()
+        >>> assert sample_vector.shape == (2,)
+        >>> assert sample_matrix.shape == (2, 3)
         >>> # xdoctest: +REQUIRES(--show)
         >>> import kwplot
         >>> kwplot.autompl()
         >>> kwplot.figure(fnum=1, pnum=(2, 2, 1))
         >>> ax = kwplot.figure(pnum=(2, 1, 1), title='n1 & n2').gca()
         >>> n = 10000
-        >>> plotkw = dict(stat='density', kde=1, bins=1000)
+        >>> #plotkw = dict(stat='density', kde=1, bins=1000)
         >>> plotkw = dict(stat='count', kde=1, bins=1000)
-        >>> #plotkw = dict(stat='frequency', kde=1, bins='auto')
         >>> n1.plot(n, ax=ax, **plotkw)
         >>> n2.plot(n, ax=ax, **plotkw)
         >>> ax=kwplot.figure(pnum=(2, 2, 3), title='composed').gca()
@@ -841,7 +847,7 @@ class Mixture(MixedDistribution):
         # Choose which distributions are picked for each sample
         idxs = self._idx_choice.sample(*shape)
         idx_to_nsamples = ub.dict_hist(idxs.ravel())
-        out = np.zeros(*shape)
+        out = np.zeros(shape)
         for idx, n in idx_to_nsamples.items():
             # Sample the from the distribution we picked
             mask = (idx == idxs)
@@ -919,6 +925,7 @@ class Composed(MixedDistribution):
         >>> kwplot.autompl()
         >>> kwplot.figure(fnum=1, doclf=True)
         >>> self.plot(1000, bins=100)
+        >>> kwplot.show_if_requested()
 
     Example:
         >>> # Binary operations result in composed distributions
@@ -935,6 +942,7 @@ class Composed(MixedDistribution):
         >>> kwplot.autompl()
         >>> kwplot.figure(fnum=1, doclf=True)
         >>> self.plot(5000, bins=100)
+        >>> kwplot.show_if_requested()
 
     """
     __params__ = ub.odict([
@@ -1038,6 +1046,7 @@ class Exponential(ContinuousDistribution):
         >>> kwplot.autompl()
         >>> kwplot.figure(fnum=1, doclf=True)
         >>> self.plot(500, bins=25)
+        >>> kwplot.show_if_requested()
 
     Args:
         scale (int) : no help given. Defaults to 1.
@@ -1134,6 +1143,7 @@ class Normal(ContinuousDistribution):
         >>> kwplot.autompl()
         >>> kwplot.figure(fnum=1, doclf=True)
         >>> self.plot(500, bins=25)
+        >>> kwplot.show_if_requested()
     """
     __params__ = ub.odict([
         ('mean', Value(0.0)),
@@ -1461,6 +1471,7 @@ class PDF(Distribution):
         >>> kwplot.autompl()
         >>> kwplot.figure(fnum=1, doclf=True)
         >>> self.plot(5000, bins=50)
+        >>> kwplot.show_if_requested()
     """
     def __init__(self, x, p, rng=None):
         import scipy.interpolate as interpolate